Allow for file handles to stay open until close() when TILEDB_KEEP_FILE_HANDLES_OPEN is set. Also, set up benchmark harness. (#56)
* Allow for file handles to also stay open until close() when TILEDB_DISABLE_FILE_LOCKING=1

* Add benchmarking harness and support TILEDB_KEEP_FILE_HANDLES_OPEN env instead of piggy backing this option through TILEDB_DISABLE_FILE_LOCKS

* Cleanup compiler warnings

+++ test/src/c_api/test_tiledb_utils.cc
@@ -172,7 +172,7 @@ TEST_CASE_METHOD(TempDir, "Test multithreaded file utils", "[file_utils_multi_th
     threads.push_back(std::move(thread_object));
   }
 
-  CHECK(num_threads == threads.size());
+  CHECK((size_t)num_threads == threads.size());
 
   for (auto i=0; i<num_threads; i++) {
     threads[i].join();
